+++ NEWS.md
@@ -1,7 +1,12 @@
+# fusionchartsR 1.2.0
+
+* Add `fusionGrid` to render interactive table
+* Upgrade version of FusionCharts Suite XT to 4.2.0 including new themes like `fint` and `access`
+* Refresh shinyApp (#11)
+
 # fusionchartsR 1.1.1
 
 * Remove `docs` folder to avoid confusion with pkgdown
-
 
 # fusionchartsR 1.1.0